backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-dotenv>=1.0.0
# Google Auth & Gmail API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
# RAG & LLM
groq>=0.4.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
# Document Processing
python-docx>=1.1.0
pypdf>=3.17.0
beautifulsoup4>=4.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
